17615 - default input parameters
In the past default input parameters had to be specified in the agent
implementation code.

Default values for input parameters can now be specified in the agent
ddl. If a rpc request is made without specifying an input value, the
input will be taken from the default value and added to the request.

Example :
   input :someinput,
         :default     => "value",
         :prompt      => "Some kind of input",
         ...

Any requests made without an input value wil default to 'value'