20059 - Data matcher logs bad errors when query returns nil
In the past when a data query field did not get assigned a value it
will default to nil which would log an non descriptive error when it gets
compared.

Data query fields that default to nil will now immediately return false,
meaning that all unset query values will evaluate to false.