20550 - allow activemq ssl certificate and key to be set via environment variables
This commit adds the ability to specify the ssl key and cert used by the
ActiveMQ and RabbitMQ connectors as evironment variables.

When enabling SSL for the rabbitmq or activemq connector plugins, the
client will first check to see if an environment variable like

  MCOLLECTIVE_ACTIVEMQ_POOLX_SSL_KEY
    or
  MCOLLECTIVE_RABBITMQ_POOLX_SSL_CERT

has been set before looking in the client config file to determine the
key and or cert.